+++ src/line.rs
@@ -1,4 +1,5 @@
 use crate::linear::Vector;
+use crate::Float;
 
 /// A structure describing a line between 2 points
 pub struct Line(pub Vector, pub Vector);
@@ -9,7 +10,7 @@ impl Line
     /// 0 is on line
     /// -1 is on one side
     /// 1 is on the other
-    pub fn edge_function(&self, point : Vector) -> f64
+    pub fn edge_function(&self, point : Vector) -> Float
     {
         (point.x - self.0.x) * (self.1.y - self.0.y) - (point.y - self.0.y) * (self.1.x - self.0.x)
     }
