+++ src/radians.rs
@@ -1,13 +1,13 @@
-use std::f64::consts::PI;
+use crate::{Float, PI};
 
 /// degrees to radians
-pub fn to_radians(deg : f64) -> f64
+pub fn to_radians(deg : Float) -> Float
 {
     deg * (PI / 180.)
 }
 
 /// radians to degrees
-pub fn to_degrees(rad : f64) -> f64
+pub fn to_degrees(rad : Float) -> Float
 {
     rad * (180. / PI)
 }
